+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * ArcEllipse2D.discretization_points()
 * ArcEllipse2D.reverse()
 * FullArc3D.linesegment_intersections()
+* ArcEllipse3D.discretization_points()
 
 
 ### Fixed
@@ -81,6 +82,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * ArcEllipse2D.discretization_points()
 * ArcEllipse2D.reverse()
 * FullArc3D.linesegment_intersections()
+* ArcEllipse3D.discretization_points()
 
 
 ## v0.6.0 [11/7/2022]
@@ -124,6 +126,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * BSplineCurve3D: from_step
 * Surface2D: cut_by_line
 * Circle3D: to_step
+* ArcEllipse3D.to_2d()
 
 
 ### Performance improvements